[+] add : job form step one component
