+++ components/common/JobFormStepTwo.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const JobFormStepTwo = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default JobFormStepTwo
